Add documentation.zip to project
This zipped file is a directory containing the documentation submitted for the C964 Capstone. The files are screenshots showing how to install Python on Windows, the raw song dataset downloaded from Kaggle.com, and a report detailing all aspects of the project (such as scenario, summaries, implementation, setup, and running the application).